Update test to verify sale price currency is made off of base price
The sales_price is made from an price associated with the variant. To verify
that the sales_price recieved the properties of the original price,
you would want to test against the original price properties. In 3.1,
this test failed due to the variant having more than one price, a master
price and a test set price. The previous behavior would ask for the last
price, `variant.prices.last.currency`, assuming a new price was generated
when it was actually pulling the master variant price which was in "USD".
In tests for 3.0 and below, the first price is also the last price,
which is why it never failed in the first place.